+++ tests/test_sphinx.py
@@ -303,7 +303,10 @@ def test_sphinx_deprecated_class_method__warns(sphinx_deprecated_class_method):
 
 def test_should_raise_type_error():
     try:
-        deprecated.sphinx.deprecated(5)
+        @deprecated.sphinx.deprecated(version="4.5.6", reason=5)
+        def foo():
+            pass
+
         assert False, "TypeError not raised"
     except TypeError:
         pass